requirements.txt
flask==2.0.2
flask-sqlalchemy==2.5.1
flask-login==0.5.0
flask-migrate===3.1.0
gunicorn==20.1.0
gevent==21.12.0
flask-wtf==1.0.0